Move downloading code to new file
